# gateway/docs/swagger.yaml
basePath: /api/gateway/
definitions:
  internal_server_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_server_auth.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user_id:
        type: integer
    type: object
  internal_server_auth.RegisterRequest:
    properties:
      birth_date:
        type: string
      email:
        type: string
      fathers_name:
        maxLength: 50
        type: string
      name:
        maxLength: 50
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
      surname:
        maxLength: 50
        type: string
    required:
    - email
    - name
    - password
    - phone
    type: object
  internal_server_auth.RegisterResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user_id:
        type: integer
    type: object
  internal_server_cart.AddToCartRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
host: localhost:7970
info:
  contact:
    email: lev.uy@mail.ru
    name: Lev U.
    url: https://github.com/LevYur/project.git
  description: Gateway service validates and routes incoming HTTP requests,
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gateway API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access & refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/internal_server_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server_auth.LoginResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access & refresh tokens
      parameters:
      - description: Register credentials
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/internal_server_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server_auth.RegisterResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User register
      tags:
      - auth
  /cart:
    delete:
      consumes:
      - application/json
      description: Clear all items from user's cart by user ID
      produces:
      - application/json
      responses:
        "200":
          description: 'success: true'
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Permission denied
        "404":
          description: Not found
        "408":
          description: Timeout
        "500":
          description: Internal error
        "503":
          description: Service unavailable
      summary: Clear all items from user's cart
      tags:
      - cart
    get:
      consumes:
      - application/json
      description: Get all items from user's cart by user ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Permission denied
        "404":
          description: Not found
        "408":
          description: Timeout
        "500":
          description: Internal error
        "503":
          description: Service unavailable
      summary: Get all items from user's cart
      tags:
      - cart
  /cart/add:
    post:
      consumes:
      - application/json
      description: Add item to user's cart by product ID + quantity
      parameters:
      - description: Add to cart payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_server_cart.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'success: true'
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Permission denied
        "404":
          description: Not found
        "408":
          description: Timeout
        "500":
          description: Internal error
        "503":
          description: Service unavailable
      summary: Add item to user's cart
      tags:
      - cart
  /cart/delete/{product_id}:
    delete:
      consumes:
      - application/json
      description: Remove item from user's cart by product ID. Removes all quantity
        of that item
      parameters:
      - description: Product ID to remove
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'success: true'
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Permission denied
        "404":
          description: Not found
        "408":
          description: Timeout
        "500":
          description: Internal error
        "503":
          description: Service unavailable
      summary: Remove item from user's cart
      tags:
      - cart
swagger: "2.0"